Refactor mfrac, mo, and mpadded into common and chtml components
