+++ mathjax3-ts/output/chtml/Wrappers/mpadded.ts
@@ -21,11 +21,9 @@
  * @author dpvc@mathjax.org (Davide Cervone)
  */
 
-import {CHTMLWrapper, StringMap} from '../Wrapper.js';
-import {BBox} from '../BBox.js';
+import {CHTMLWrapper, CHTMLConstructor, StringMap} from '../Wrapper.js';
+import {CommonMpadded, CommonMpaddedInterface} from '../../common/Wrappers/mpadded.js';
 import {MmlMpadded} from '../../../core/MmlTree/MmlNodes/mpadded.js';
-import {MmlNode} from '../../../core/MmlTree/MmlNode.js';
-import {Property} from '../../../core/Tree/Node.js';
 import {StyleList} from '../../common/CssStyles.js';
 
 /*****************************************************************/
@@ -36,7 +34,8 @@ import {StyleList} from '../../common/CssStyles.js';
  * @template T  The Text node class
  * @template D  The Document class
  */
-export class CHTMLmpadded<N, T, D> extends CHTMLWrapper<N, T, D> {
+export class CHTMLmpadded<N, T, D> extends CommonMpadded<N, T, D, CHTMLConstructor<N, T, D>>(CHTMLWrapper) {
+
     public static kind = MmlMpadded.prototype.kind;
 
     public static styles: StyleList = {
@@ -84,60 +83,6 @@ export class CHTMLmpadded<N, T, D> extends CHTMLWrapper<N, T, D> {
         for (const child of this.childNodes) {
             child.toCHTML(content[0] || chtml);
         }
-    }
-
-    /**
-     * Get the content bounding box, and the change in size and offsets
-     *   as specified by the parameters
-     *
-     * @return {number[]}  The original height, depth, width, the changes in height, depth,
-     *                    and width, and the horizontal and vertical offsets of the content
-     */
-    protected getDimens() {
-        const values = this.node.attributes.getList('width', 'height', 'depth', 'lspace', 'voffset');
-        const bbox = this.childNodes[0].getBBox();  // get unmodified bbox of children
-        let {w, h, d} = bbox;
-        let W = w, H = h, D = d, x = 0, y = 0;
-        if (values.width !== '')   w = this.dimen(values.width, bbox, 'w', 0);
-        if (values.height !== '')  h = this.dimen(values.height, bbox, 'h', 0);
-        if (values.depth !== '')   d = this.dimen(values.depth, bbox, 'd', 0);
-        if (values.voffset !== '') y = this.dimen(values.voffset, bbox);
-        if (values.lspace !== '')  x = this.dimen(values.lspace, bbox);
-        return [H, D, W, h - H, d - D, w - W, x, y];
-    }
-
-    /**
-     * Get a particular dimension, which can be relative to any of the BBox dimensions,
-     *   and can be an offset from the default size of the given dimension.
-     *
-     * @param {Property} length   The value to be converted to a length in ems
-     * @param {BBox} bbox         The bbox of the mpadded content
-     * @param {string} d          The default dimension to use for relative sizes ('w', 'h', or 'd')
-     * @param {number} m          The minimum value allowed for the dimension
-     * @return {number}           The final dimension in ems
-     */
-    protected dimen(length: Property, bbox: BBox, d: string = '', m: number = null) {
-        length = String(length);
-        const match = length.match(/width|height|depth/);
-        const size = (match ? bbox[match[0].charAt(0) as (keyof BBox)] : (d ? bbox[d as (keyof BBox)] : 0)) as number;
-        let dimen = (this.length2em(length, size) || 0);
-        if (length.match(/^[-+]/) && d) {
-            dimen += size;
-        }
-        if (m != null) {
-            dimen = Math.max(m, dimen);
-        }
-        return dimen;
-    }
-
-    /**
-     * @override
-     */
-    public computeBBox(bbox: BBox) {
-        const [H, D, W, dh, dd, dw, x, y] = this.getDimens();
-        bbox.w = W + dw;
-        bbox.h = H + dh;
-        bbox.d = D + dd;
     }
 
 }
